+++ app/src/main/kotlin/com/konifar/materialcat/infra/di/RepositoryModule.kt
@@ -5,14 +5,12 @@ import com.konifar.materialcat.infra.repository.catphoto.CatImageRepository
 import com.konifar.materialcat.infra.repository.catphoto.FlickrCatImageRepositoryImpl
 import dagger.Module
 import dagger.Provides
-import javax.inject.Named
 import javax.inject.Singleton
 
 
 @Module
 class RepositoryModule() {
 
-    @Named("flickrCatImageRepository")
     @Singleton
     @Provides
     fun provideFlickrCatImageRepository(flickrApiService: FlickrApiService): CatImageRepository = FlickrCatImageRepositoryImpl(flickrApiService)
