+++ app/src/main/kotlin/com/konifar/materialcat/presentation/di/ActivityComponent.kt
@@ -5,7 +5,7 @@ import com.konifar.materialcat.PhotoDetailActivity
 import dagger.Subcomponent
 
 @ActivityScope
-@Subcomponent(modules = arrayOf(ActivityModule::class, UseCaseModule::class))
+@Subcomponent(modules = arrayOf(ActivityModule::class))
 interface ActivityComponent {
 
     fun inject(activity: MainActivity)
